[refactor] fcm 푸시알림을 위한 이벤트 리스너 하위 코드 수정 (#296)
- 이벤트에 포함된 title, content 정보를 받아서 바로 사용하도록 수정
- title, content 의 생성 책임은 template enum 만 담당
