+++ src/main/java/konkuk/thip/message/application/service/RoomNotificationDispatchService.java
@@ -5,7 +5,7 @@ import com.google.firebase.messaging.Notification;
 import konkuk.thip.message.application.port.in.RoomNotificationDispatchUseCase;
 import konkuk.thip.message.application.port.out.FirebaseMessagingPort;
 import konkuk.thip.message.adapter.out.event.dto.RoomEvents;
-import konkuk.thip.message.domain.NotificationCategory;
+import konkuk.thip.notification.domain.value.NotificationCategory;
 import konkuk.thip.message.domain.MessageRoute;
 import konkuk.thip.notification.application.port.out.FcmTokenPersistencePort;
 import konkuk.thip.notification.domain.FcmToken;
@@ -26,8 +26,7 @@ public class RoomNotificationDispatchService implements RoomNotificationDispatch
 
     @Override
     public void handleRoomPostCommented(final RoomEvents.RoomPostCommentedEvent event) {
-        Notification notification = buildNotification("새로운 댓글이 달렸어요",
-                "@" + event.actorUsername() + " 님이 내 독서기록에 댓글을 달았어요!");
+        Notification notification = buildNotification(event.title(), event.content());
 
         List<FcmToken> tokens = fcmTokenQueryPort.findEnabledByUserId(event.targetUserId());
         if (tokens.isEmpty()) return;
@@ -52,8 +51,7 @@ public class RoomNotificationDispatchService implements RoomNotificationDispatch
 
     @Override
     public void handleRoomVoteStarted(final RoomEvents.RoomVoteStartedEvent event) {
-        Notification notification = buildNotification(event.roomTitle(),
-                "새로운 투표가 시작되었어요!");
+        Notification notification = buildNotification(event.title(), event.content());
 
         List<FcmToken> tokens = fcmTokenQueryPort.findEnabledByUserId(event.targetUserId());
         if (tokens.isEmpty()) return;
@@ -78,8 +76,7 @@ public class RoomNotificationDispatchService implements RoomNotificationDispatch
 
     @Override
     public void handleRoomRecordCreated(final RoomEvents.RoomRecordCreatedEvent event) {
-        Notification notification = buildNotification(event.roomTitle(),
-                "@" + event.actorUsername() + " 님이 새로운 독서 기록을 작성했어요!");
+        Notification notification = buildNotification(event.title(), event.content());
 
         List<FcmToken> tokens = fcmTokenQueryPort.findEnabledByUserId(event.targetUserId());
         if (tokens.isEmpty()) return;
@@ -104,32 +101,28 @@ public class RoomNotificationDispatchService implements RoomNotificationDispatch
 
     @Override
     public void handleRoomRecruitClosedEarly(final RoomEvents.RoomRecruitClosedEarlyEvent event) {
-        Notification n = buildNotification(event.roomTitle(),
-                "모임방 활동이 시작되었어요. 모임방에서 독서 기록을 시작해보세요!");
+        Notification notification = buildNotification(event.title(), event.content());
 
-        pushRoomMain(event.targetUserId(), event.roomId(), n);
+        pushRoomMain(event.targetUserId(), event.roomId(), notification);
     }
 
     @Override
     public void handleRoomActivityStarted(final RoomEvents.RoomActivityStartedEvent event) {
-        Notification notification = buildNotification(event.roomTitle(),
-                "모임방 활동이 시작되었어요. 모임방에서 독서 기록을 시작해보세요!");
+        Notification notification = buildNotification(event.title(), event.content());
 
         pushRoomMain(event.targetUserId(), event.roomId(), notification);
     }
 
     @Override
     public void handleRoomJoinRequestedToOwner(final RoomEvents.RoomJoinRequestedToOwnerEvent event) {
-        Notification n = buildNotification(event.roomTitle(),
-                "@" + event.applicantUsername() + " 님이 모임에 참여했어요!");
+        Notification notification = buildNotification(event.title(), event.content());
 
-        pushRoomDetail(event.ownerUserId(), event.roomId(), n);
+        pushRoomDetail(event.ownerUserId(), event.roomId(), notification);
     }
 
     @Override
     public void handleRoomCommentLiked(final RoomEvents.RoomCommentLikedEvent event) {
-        Notification notification = buildNotification("내 댓글을 좋아합니다",
-                "@" + event.actorUsername() + " 님이 내 댓글에 좋아요를 눌렀어요!");
+        Notification notification = buildNotification(event.title(), event.content());
 
         List<FcmToken> tokens = fcmTokenQueryPort.findEnabledByUserId(event.targetUserId());
         if (tokens.isEmpty()) return;
@@ -153,8 +146,7 @@ public class RoomNotificationDispatchService implements RoomNotificationDispatch
 
     @Override
     public void handleRoomPostLiked(final RoomEvents.RoomPostLikedEvent event) {
-        Notification notification = buildNotification("좋아요 알림",
-                "@" + event.actorUsername() + " 님이 내 독서기록에 좋아요를 눌렀어요!");
+        Notification notification = buildNotification(event.title(), event.content());
 
         List<FcmToken> tokens = fcmTokenQueryPort.findEnabledByUserId(event.targetUserId());
         if (tokens.isEmpty()) return;
@@ -177,11 +169,10 @@ public class RoomNotificationDispatchService implements RoomNotificationDispatch
     }
 
     @Override
-    public void handleRoomPostCommentReplied(RoomEvents.RoomPostCommentRepliedEvent e) {
-        Notification notification = buildNotification("새로운 댓글이 달렸어요",
-                "@" + e.actorUsername() + " 님이 내 댓글에 대댓글을 달았어요!");
+    public void handleRoomPostCommentReplied(RoomEvents.RoomPostCommentRepliedEvent event) {
+        Notification notification = buildNotification(event.title(), event.content());
 
-        List<FcmToken> tokens = fcmTokenQueryPort.findEnabledByUserId(e.targetUserId());
+        List<FcmToken> tokens = fcmTokenQueryPort.findEnabledByUserId(event.targetUserId());
         if (tokens.isEmpty()) return;
 
         List<Message> msgs = new ArrayList<>(tokens.size());
@@ -191,11 +182,11 @@ public class RoomNotificationDispatchService implements RoomNotificationDispatch
         for (FcmToken t : tokens) {
             Message m = buildMessage(t.getFcmToken(), notification,
                     MessageRoute.ROOM_POST_DETAIL,
-                    "roomId", String.valueOf(e.roomId()),
-                    "page", String.valueOf(e.page()),
+                    "roomId", String.valueOf(event.roomId()),
+                    "page", String.valueOf(event.page()),
                     "type", "group",
-                    "postId", String.valueOf(e.postId()),
-                    "postType", String.valueOf(e.postType()));
+                    "postId", String.valueOf(event.postId()),
+                    "postType", String.valueOf(event.postType()));
 
             msgs.add(m); tk.add(t.getFcmToken()); dev.add(t.getDeviceId());
         }
@@ -243,7 +234,7 @@ public class RoomNotificationDispatchService implements RoomNotificationDispatch
     }
 
     private Notification buildNotification(final String title, final String body) {
-        return Notification.builder().setTitle(NotificationCategory.ROOM.prefixedTitle(title)).setBody(body).build();
+        return Notification.builder().setTitle(title).setBody(body).build();
     }
 
     private Message buildMessage(final String token, final Notification n,
